integrado con errores parte de Juan
