+++ config.json
@@ -1,0 +1,1 @@
+{"estado_alarma": 0, "tono_alarma": "alarm", "tonos_alarma": ["alarm", "y2mate.com_-_Yamete_Kudasai_Sound_Effect_Original"], "volumen_alarma": "78"}
