+++ main.py
@@ -8,9 +8,28 @@ from object_detection.utils import visualization_utils as viz_utils
 from object_detection.builders import model_builder
 from playsound import playsound
 
-#packages for the web server
-from flask import Flask, render_template, Response
+# web server and socket
+from flask import Flask, render_template, request, jsonify, redirect, Response
+from flask_socketio import SocketIO, emit
+from werkzeug.utils import secure_filename
 
+import json, os
+
+#  variables y settings
+data = dict()
+with open('config.json') as json_file:
+    data = json.load(json_file)
+
+ALLOWED_EXTENSIONS = ['mp3']
+UPLOAD_FOLDER = os.path.join(os.path.abspath(os.getcwd()), 'static')
+UPLOAD_FOLDER = os.path.join(UPLOAD_FOLDER, 'alarms')
+
+app = Flask(__name__, template_folder='templates')
+app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+
+socketio = SocketIO(app)
+
+#methods
 def detect(image, detection_model):
     image, shapes = detection_model.preprocess(image)
     prediction_dict = detection_model.predict(image, shapes)
@@ -18,11 +37,12 @@ def detect(image, detection_model):
 
     return detections, prediction_dict, tf.reshape(shapes, [-1])
 
-
 def run(label_map_path, config_file_path, checkpoint_path):
+    global app
 
     # Enable GPU dynamic memory allocation
     gpus = tf.config.experimental.list_physical_devices('GPU')
+
     for gpu in gpus:
         tf.config.experimental.set_memory_growth(gpu, True)
 
@@ -37,7 +57,7 @@ def run(label_map_path, config_file_path, checkpoint_path):
 
     category_index = label_map_util.create_category_index_from_labelmap(label_map_path, use_display_name=True)
 
-    videoStreamAddress = "http://192.168.1.82:4747/video"
+    videoStreamAddress = "http://192.168.1.102:4747/video"
 
     cap = cv2.VideoCapture(videoStreamAddress) #esto básicamente
     
@@ -65,8 +85,7 @@ def run(label_map_path, config_file_path, checkpoint_path):
 
         label_id_offset = 1
         image_np_with_detections = image_np.copy()
-        bb_color = 'green'
-
+        border_color = 'green'
 
         if indexes.shape != 0:
             viz_utils.visualize_boxes_and_labels_on_image_array(
@@ -82,62 +101,91 @@ def run(label_map_path, config_file_path, checkpoint_path):
                 semaphore_mode=True)
 
             # si cualquiera de las personas frente a la cámara está usando mal la mascarilla o no tiene, marcar recuadro rojo
-            if any(c == 1 or c == 3 for c in detection_classes):
-                bb_color = 'green' #rojo
+            if any(c == 0 or c == 2 for c in detection_classes):
+                border_color = 'blue' #rojo (bug de TF)
+                with app.app_context():
+                    emit('alarm', {'alarm': 1}, namespace='/', broadcast=True)
+
             # sino, todo ok, verde
             else:
-                bb_color = 'blue' #verde
+                border_color = 'green' #verde
+                with app.app_context():
+                    emit('alarm', {'alarm': 0}, namespace='/', broadcast=True)
             
             viz_utils.draw_bounding_box_on_image_array(
                 image_np_with_detections,
                 0, 0, 1, 1, 
-                color = bb_color,
+                color = border_color,
                 thickness = 20)
 
-        print(detection_boxes, detection_classes, detection_scores)
-        print((detection_classes + label_id_offset).astype(int))
-        
-        if any(c == 1 or c == 3 for c in detection_classes):
-           playsound('templates/alarm.mp3')
-
         frame = cv2.resize(image_np_with_detections, (800, 600))
-
-        # Display output
-        #cv2.imshow('MaskON: Uso correcto de mascarilla', cv2.resize(image_np_with_detections, (800, 600)))
-        # encode OpenCV raw frame to jpg and displaying it
         
         ret, jpeg = cv2.imencode('.jpg', frame)
         frame_encoded_jpeg = jpeg.tobytes()
 
         yield (b'--frame\r\n'
                b'Content-Type: image/jpeg\r\n\r\n' + frame_encoded_jpeg + b'\r\n\r\n')
-    
+        
         if cv2.waitKey(25) & 0xFF == ord('q'):
             break
 
     cap.release()
     cv2.destroyAllWindows()
 
-#mounting web server TODO: Searar en un módulo aparte.
-app = Flask(__name__)
-
-@app.route('/')
-def index():
-    # rendering webpage
-    return render_template('index.html')
-
-@app.route('/video_feed')
-def video_feed():
+# RUTAS
+@app.route('/cam')
+def cam():
     return Response(run(
                         label_map_path=PATH_TO_LABELMAP,
                         config_file_path=PATH_TO_CONFIG,
                         checkpoint_path=PATH_TO_CHECKPOINT),
                     mimetype='multipart/x-mixed-replace; boundary=frame')
 
-if __name__ == '__main__':
-    PATH_TO_LABELMAP = r"C:\Users\matia\Documents\Scripts\PDI\training_resources\dataset\label_map.pbtxt"
-    PATH_TO_CONFIG = r"C:\Users\matia\Documents\Scripts\PDI\training_resources\training\pipeline.config"
-    PATH_TO_CHECKPOINT = r"C:\Users\matia\Documents\Scripts\PDI\training_resources\training\003\ckpt-16"
+@app.route('/')
+def root():
+    return redirect("index", code=303)
 
-    #run(label_map_path=PATH_TO_LABELMAP, config_file_path=PATH_TO_CONFIG, checkpoint_path=PATH_TO_CHECKPOINT)
-    app.run(host='127.0.0.1',port='5000', debug=True)
+@app.route('/index')
+def index():
+    resp = render_template("index.html")
+    return resp
+
+@app.route('/settings')
+def settings():
+    resp = render_template("settings.html")
+    shutdown = True
+    return resp
+
+
+@app.route('/get_config')
+def get_settings():
+    return jsonify(data)
+
+@app.route('/post_config', methods=['POST'])
+def post_settings():
+    global data
+    if request.method == 'POST':
+        data = request.get_json()
+        with open('config.json', 'w') as fp:
+            json.dump(data, fp)
+    return '200'
+
+@app.route('/post_file', methods=['POST'])
+def post_file():
+    f = request.files['file']
+    print(os.path.join(app.config['UPLOAD_FOLDER'], f.filename))
+    f.save(os.path.join(app.config['UPLOAD_FOLDER'], secure_filename(f.filename)))
+    return '200'
+
+@socketio.on('alarm_change')
+def alarm_event(json):
+    print(json)
+    emit('my response', json, broadcast=True)
+
+if __name__ == '__main__':
+    PATH_TO_LABELMAP = r"C:\Users\matia\Documents\Scripts\PDI\training_resources\training\001\label_map.pbtxt"
+    PATH_TO_CONFIG = r"C:\Users\matia\Documents\Scripts\PDI\training_resources\training\pipeline.config"
+    PATH_TO_CHECKPOINT = r"C:\Users\matia\Documents\Scripts\PDI\training_resources\training\001\ckpt-16"
+
+    #app.run(host='127.0.0.1',port='5000', debug=True)
+    socketio.run(app)
