+++ object_detection/utils/visualization_utils.py
@@ -44,9 +44,11 @@ from object_detection.utils import shape_utils
 
 _TITLE_LEFT_MARGIN = 10
 _TITLE_TOP_MARGIN = 10
+
 SEMAPHORE_COLORS= [
-  'blue', 'green', '#ffff00' 
+  'blue', 'green', 'yellow' 
 ]
+
 STANDARD_COLORS = [
     'AliceBlue', 'Chartreuse', 'Aqua', 'Aquamarine', 'Azure', 'Beige', 'Bisque',
     'BlanchedAlmond', 'BlueViolet', 'BurlyWood', 'CadetBlue', 'AntiqueWhite',
