+++ static/js/index.js
@@ -1,7 +1,6 @@
 var data = {};
 var audio = null;
 
-
 function configure(){
     var sonido_alarma = data['tono_alarma']; //EXTRAER DE SESSIONSTORAGE
     audio = new Audio('/static/alarms/'+sonido_alarma+'.mp3'); //PONER NOMBRE EN B
@@ -14,20 +13,19 @@ function check_data(){
         //aqui se hace el get al servidor
         // se guarga en session storage
         $.ajax({
-        url: "/get_config",
-        type: 'GET',
-        success: function(res) {
-            sessionStorage.setItem('data',JSON.stringify(res))
-            data=JSON.parse(sessionStorage.getItem('data'));
-            configure();
-        }
-    });
+            url: "/get_config",
+            type: 'GET',
+            success: function(res) {
+                sessionStorage.setItem('data',JSON.stringify(res))
+                data=JSON.parse(sessionStorage.getItem('data'));
+                configure();
+            }
+        });
     }
-    else{
+    else {
         configure();
     }
 }
-
 
 function post_data(){
     sessionStorage.setItem('data',JSON.stringify(data))
@@ -41,12 +39,11 @@ function post_data(){
             console.log(res);
 
         }
-    })
+    });
 }
 
-
 function alarm(){
-    if(data['estado_alarma']==1){
+    if(data['estado_alarma'] == 1){
         audio.loop=true;
         audio.play();
     }
@@ -58,22 +55,22 @@ function stop_alarm(){
 }
 
 function mute_alarm(){
-    if(current_sound==1){
+    if (current_sound == 1){
         audio.pause();
         audio.currentTime = 0;
     }
 
-    if(data['estado_alarma']==1){
-       data['estado_alarma']=0;
+    if (data['estado_alarma'] == 1){
+       data['estado_alarma'] = 0;
        post_data();
        document.getElementById("mute_button").innerHTML = 'Quitar Silencio';
 
     }
-    else if(data['estado_alarma']==0){
-       data['estado_alarma']=1;
+    else if (data['estado_alarma'] == 0){
+       data['estado_alarma'] = 1;
        post_data();
        document.getElementById("mute_button").innerHTML = "Silenciar";
-        if(current_sound==1){
+        if (current_sound == 1){
             alarm();
         }
     }
