+++ templates/index.html
@@ -1,9 +1,57 @@
+<!DOCTYPE html>
 <html>
-  <head>
-    <title>MaskON: Uso correcto de mascarilla</title>
-  </head>
-  <body>
-    <h1>MaskON: Uso correcto de mascarilla</h1>
-    <img id="bg" src="{{ url_for('video_feed') }}">
-  </body>
+<head>
+	<title>MaskON: Uso correcto de mascarilla</title>
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link rel="stylesheet" href="{{ url_for('static', filename='css/index.css') }}">
+	<script src="{{url_for('static', filename='js/index.js')}}"></script>
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
+	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+	<script src="//cdnjs.cloudflare.com/ajax/libs/socket.io/2.2.0/socket.io.js" integrity="sha256-yr4fRk/GU1ehYJPAs8P4JlTgu0Hdsp4ZKrx8bDEDC3I=" crossorigin="anonymous"></script>
+
+
+	<script>
+		var socket = io();
+		var current_sound = 2;
+
+    	socket.on('connect', function() {
+			socket.emit('alarm_change', {data: 'I\'m connected from home!'});
+    	});
+
+    	socket.on('alarm', function(data) {
+			if(data['alarm'] == 1 && data['alarm'] != current_sound){
+				console.log('alarmando')
+				current_sound = 1;
+				alarm()
+			}
+			else if(data['alarm'] == 0 && data['alarm'] != current_sound){
+				console.log('desalarmando')
+				current_sound = 0;
+				stop_alarm()
+			}
+    	});
+	</script>
+
+
+</head>
+<body style="background-color: gray" onload="check_data()">
+	<div class="container">
+		<div style="padding-top: 100px; color:white">
+			<h1>MaskON: Uso correcto de mascarilla</h1>
+		</div>
+		<div id="video_stream"  ><img id="bg" src="{{ url_for('cam') }}"></div>
+	</div>
+</body>
+<footer class="container">
+	<div class="row">
+		<div class="col-1"></div>
+		<div class="col-2"><button type="button" class="btn btn-primary" id="mute_button" onclick="mute_alarm()">Silenciar alarma</button></div>
+		<div class="col-6"></div>
+		<div class="col-2"><button type="button" class="btn btn-primary" onclick="location.href=&quot;{{ url_for('settings') }}&quot;">Configuraciones</button></div>
+		<div class="col-1"></div>
+	</div>
+</footer>
 </html>
