+++ templates/settings.html
@@ -21,7 +21,7 @@
 
 	<title>hola!</title>
 </head>
-<body style="background-color:darkblue" onload="check_data()">
+<body style="background-color:grey" onload="check_data()">
 	<div style="padding-top: 50px"></div>
 
 	<div class="container" style="background-color: rgba(240,248,255,0.5); vertical-align:center">
