--- web/config.json
@@ -1,1 +1,0 @@
-{"estado_alarma": 1, "tono_alarma": "alarm", "tonos_alarma": ["alarm", "y2mate.com_-_Yamete_Kudasai_Sound_Effect_Original"], "volumen_alarma": "39"}
