--- web/requirements.txt
@@ -1,23 +1,0 @@
-bidict==0.21.2
-certifi==2020.12.5
-cffi==1.14.4
-click==7.1.2
-dnspython==1.16.0
-eventlet==0.30.0
-Flask==1.1.2
-Flask-SocketIO==4.3.2
-gevent==20.12.1
-greenlet==0.4.17
-itsdangerous==1.1.0
-Jinja2==2.11.2
-MarkupSafe==1.1.1
-pycparser==2.20
-python-engineio==3.13.2
-python-socketio==4.6.0
-six==1.15.0
-style==1.1.0
-update==0.0.1
-Werkzeug==1.0.1
-wincertstore==0.2
-zope.event==4.5.0
-zope.interface==5.2.0
