--- web/requirements_old.txt
@@ -1,6 +1,0 @@
-flask
-eventlet
-gevent
-python-engineio==3.13.2
-python-socketio==4.6.0
-flask-socketio==4.3.2
